chore: move faq section to FAQ component
