+++ src/pages/Contact.tsx
@@ -1,115 +1,20 @@
-import React, { useState, useEffect, useRef } from 'react';
+import React from 'react';
 import {
   ContactWrapper,
   MainContent,
-  Section,
-  Title,
-  FormWrapper,
-  Input,
-  TextArea,
-  SubmitButton,
   TeamImages,
   TeamImage,
   ReplyText,
-  FAQSection,
-  FAQTitle,
-  FAQGrid,
-  FAQItem,
-  Question,
-  Answer,
-  Separator,
-  StyledImage,
 } from '../styles/ContactStyles';
 import Header from '../components/Header';
 import Footer from '../components/Footer';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import {
-  faChevronRight,
-  faChevronUp,
-} from '@fortawesome/free-solid-svg-icons';
 import MarquesZahir from '../assets/Marques-Zahir.png'; // Replace with actual image path
 import CaseyBass from '../assets/Casey-Bass.png'; // Replace with actual image path
 import AshleyLewis from '../assets/Ashley-Lewis.png'; // Replace with actual image path
 import connectTeam from '../assets/connect-team.webp'; // Replace with actual image path
-import faqData from '../constants/faq';
+import FAQ from '../components/contact/FAQ';
 
 const Contact: React.FC = () => {
-  const [activeIndexes, setActiveIndexes] = useState<{
-    [key: number]: number | null;
-  }>({});
-  const faqGridRef = useRef<HTMLDivElement>(null);
-
-  const handleToggle = (
-    itemIndex: number,
-    questionIndex: number
-  ) => {
-    setActiveIndexes((prev) => ({
-      ...prev,
-      [itemIndex]:
-        prev[itemIndex] === questionIndex
-          ? null
-          : questionIndex,
-    }));
-  };
-
-  useEffect(() => {
-    const faqGrid = faqGridRef.current;
-    if (!faqGrid) return;
-
-    let isDown = false;
-    let startX: number;
-    let scrollLeft: number;
-
-    const handleMouseDown = (e: MouseEvent) => {
-      isDown = true;
-      faqGrid.classList.add('active');
-      startX = e.pageX - faqGrid.offsetLeft;
-      scrollLeft = faqGrid.scrollLeft;
-    };
-
-    const handleMouseLeave = () => {
-      isDown = false;
-      faqGrid.classList.remove('active');
-    };
-
-    const handleMouseUp = () => {
-      isDown = false;
-      faqGrid.classList.remove('active');
-    };
-
-    const handleMouseMove = (e: MouseEvent) => {
-      if (!isDown) return;
-      e.preventDefault();
-      const x = e.pageX - faqGrid.offsetLeft;
-      const walk = (x - startX) * 3; //scroll-fast
-      faqGrid.scrollLeft = scrollLeft - walk;
-    };
-
-    faqGrid.addEventListener('mousedown', handleMouseDown);
-    faqGrid.addEventListener(
-      'mouseleave',
-      handleMouseLeave
-    );
-    faqGrid.addEventListener('mouseup', handleMouseUp);
-    faqGrid.addEventListener('mousemove', handleMouseMove);
-
-    return () => {
-      faqGrid.removeEventListener(
-        'mousedown',
-        handleMouseDown
-      );
-      faqGrid.removeEventListener(
-        'mouseleave',
-        handleMouseLeave
-      );
-      faqGrid.removeEventListener('mouseup', handleMouseUp);
-      faqGrid.removeEventListener(
-        'mousemove',
-        handleMouseMove
-      );
-    };
-  }, []);
-
   return (
     <>
       <Header />
@@ -129,10 +34,10 @@ const Contact: React.FC = () => {
               />
             </div>
             <div className="bg-[#19a530] w-full rounded-[48px]">
-              TEST
+              TESTReplace with actual image path
             </div>
           </div>
-          {/* <TeamImages>
+          <TeamImages>
             <TeamImage
               src={MarquesZahir}
               alt="Marques Zahir"
@@ -145,58 +50,8 @@ const Contact: React.FC = () => {
             <ReplyText>
               Hang Tight, We'll Reply Shortly.
             </ReplyText>
-          </TeamImages> */}
-          <FAQSection>
-            <FAQTitle>FAQ</FAQTitle>
-            <FAQGrid ref={faqGridRef}>
-              {faqData.map((item, itemIndex) => (
-                <FAQItem key={itemIndex}>
-                  {item.questions.map(
-                    (question, questionIndex) => (
-                      <div key={questionIndex}>
-                        <Question
-                          className={
-                            activeIndexes[itemIndex] ===
-                            questionIndex
-                              ? 'active'
-                              : ''
-                          }
-                          onClick={() =>
-                            handleToggle(
-                              itemIndex,
-                              questionIndex
-                            )
-                          }
-                        >
-                          {question}{' '}
-                          <FontAwesomeIcon
-                            icon={
-                              activeIndexes[itemIndex] ===
-                              questionIndex
-                                ? faChevronUp
-                                : faChevronRight
-                            }
-                          />
-                        </Question>
-                        <Answer
-                          className={
-                            activeIndexes[itemIndex] ===
-                            questionIndex
-                              ? 'active'
-                              : ''
-                          }
-                        >
-                          {item.answers[questionIndex]}
-                        </Answer>
-                        {activeIndexes[itemIndex] !==
-                          questionIndex && <Separator />}
-                      </div>
-                    )
-                  )}
-                </FAQItem>
-              ))}
-            </FAQGrid>
-          </FAQSection>
+          </TeamImages>
+          <FAQ />
         </MainContent>
       </ContactWrapper>
       <Footer />
